chore(query): merge group_by into aggregate
