+++ src/query/service/src/pipelines/processors/transforms/aggregator/serde/mod.rs
@@ -18,9 +18,6 @@ mod transform_aggregate_spill_writer;
 mod transform_deserializer;
 mod transform_exchange_aggregate_serializer;
 mod transform_exchange_async_barrier;
-mod transform_exchange_group_by_serializer;
-mod transform_group_by_serializer;
-mod transform_group_by_spill_writer;
 mod transform_spill_reader;
 
 pub use serde_meta::*;
@@ -29,9 +26,6 @@ pub use transform_aggregate_spill_writer::*;
 pub use transform_deserializer::*;
 pub use transform_exchange_aggregate_serializer::*;
 pub use transform_exchange_async_barrier::*;
-pub use transform_exchange_group_by_serializer::*;
-pub use transform_group_by_serializer::*;
-pub use transform_group_by_spill_writer::*;
 pub use transform_spill_reader::*;
 
 pub mod exchange_defines {
