+++ src/query/service/src/pipelines/processors/transforms/aggregator/serde/transform_deserializer.rs
@@ -181,5 +181,4 @@ impl BlockMetaTransform<ExchangeDeserializeMeta> for TransformDeserializer {
     }
 }
 
-pub type TransformGroupByDeserializer = TransformDeserializer;
 pub type TransformAggregateDeserializer = TransformDeserializer;
