+++ src/query/service/src/pipelines/processors/transforms/aggregator/serde/transform_spill_reader.rs
@@ -311,5 +311,4 @@ impl TransformSpillReader {
     }
 }
 
-pub type TransformGroupBySpillReader = TransformSpillReader;
 pub type TransformAggregateSpillReader = TransformSpillReader;
